<< implementado
